+++ plotlykt-server/src/main/kotlin/space/kscience/plotly/server/Update.kt
@@ -1,10 +1,15 @@
 package space.kscience.plotly.server
 
+import kotlinx.serialization.json.JsonArray
 import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.put
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.meta.toJson
+
+
+private val coordinateNames  =
+    listOf("x", "y", "z", "text", "close", "high", "low", "open")
 
 /**
  * An update message for both data and layout
@@ -17,9 +22,13 @@ public sealed class Update(public val id: String) {
             put("plotId", id)
             put("contentType", "trace")
             put("trace", trace)
-            put("content", content.toJson())
+            //patch json to adhere to plotly array in array specification
+            val contentJson = content.toJson()
+            val patchedJson = contentJson + coordinateNames.associateWith { contentJson[it] }
+                .filter { it.value != null }
+                .mapValues { JsonArray(listOf(it.value!!)) }
+            put("content", JsonObject(patchedJson))
         }
-
     }
 
     public class Layout(id: String, private val content: Meta) : Update(id) {